feat: allow passing additional props to pager
In cases when TabView is used along with another handler of react-native-gesture-handler, it's required to pass additional props to underlying PanGestureHandler, such as various refs to wait for/simultaneously work with etc. Different people need to pass different props, so I decided to accept and spread just an object of props instead of listing each of "allowed" additional props. There are no Flow typedefs in react-native-gesture-handler, so the type of pagerProps is just Object.